fix: refactor fns
